+++ buttons/buttons_playlist_history.js
@@ -1,0 +1,40 @@
+﻿'use strict';
+
+/* 
+	Playlist History
+	----------------
+	Switch to previous playlists.
+ */
+
+include(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\helpers\\buttons_xxx.js'); 
+try { //May be loaded along other buttons
+	window.DefinePanel('Playlist Tools Menu', {author:'xxx'});
+	var g_font = _gdiFont('Segoe UI', 12);
+	var buttonCoordinates = {x: 0, y: 0, w: 98, h: 22};
+	var buttonOrientation = 'x';
+} catch (e) {
+	buttonCoordinates = {x: 0, y: 0, w: buttonOrientation === 'x' ? 98 : buttonCoordinates.w , h: buttonOrientation === 'y' ? 22 : buttonCoordinates.h}; // Reset 
+	console.log('Playlist Tools Menu Button loaded.');
+}
+include(fb.ProfilePath + 'scripts\\SMP\\xxx-scripts\\helpers\\playlist_history.js');
+
+var newButtons = {
+	menuButton: new SimpleButton(calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? true : false).x, calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? false : true).y, buttonCoordinates.w, buttonCoordinates.h, "Prev. Playlist", function (mask) {
+		if (mask === MK_SHIFT) {
+			createHistoryMenu().btn_up(this.x, this.y + this.h);
+		} else {
+			goPrevPls();
+		}
+	}, null, g_font, () => {return 'Switch to previous playlist:\n' + getPrevPlsName();}, null, null, chars.prevCircle, _gdiFont('FontAwesome', 10)),
+};
+// Check if the button list already has the same button ID
+for (var buttonName in newButtons) {
+	if (buttons.hasOwnProperty(buttonName)) {
+		// fb.ShowPopupMessage('Duplicated button ID (' + buttonName + ') on ' + window.Name);
+		console.log('Duplicated button ID (' + buttonName + ') on ' + window.Name);
+		Object.defineProperty(newButtons, buttonName + Object.keys(buttons).length, Object.getOwnPropertyDescriptor(newButtons, buttonName));
+		delete newButtons[buttonName];
+	}
+}
+// Adds to current buttons
+buttons = {...buttons, ...newButtons};