+++ buttons/buttons_remove_duplicates.js
@@ -43,7 +43,7 @@ setProperties(newButtonsProperties, prefix); //This sets all the panel propertie
 if (buttonOrientation === 'x') {buttonCoordinates.w += 25;}
 
 var newButtons = {
-	RemoveDuplicates: new SimpleButton(calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? true : false).x, calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? false : true).y, buttonCoordinates.w, buttonCoordinates.h, 'Remove Duplicates', function () {
+	RemoveDuplicates: new SimpleButton(calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? true : false).x, calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? false : true).y, buttonCoordinates.w, buttonCoordinates.h, 'Rmv. duplicates', function () {
 		let t0 = Date.now();
 		let t1 = 0;
 		let badSortInput = getPropertiesValues(this.buttonsProperties, this.prefix, void(0), 4); //This gets all the panel properties at once but 4th
@@ -51,7 +51,7 @@ var newButtons = {
         do_remove_duplicatesV2(null, null, sortInput);
 		t1 = Date.now();
 		console.log("Call to do_remove_duplicatesV2 took " + (t1 - t0) + " milliseconds.");
-	}, null, g_font,'Removes duplicates according to equal ' + enumeratePropertiesValues(newButtonsProperties, prefix, void(0), void(0), 4), prefix, newButtonsProperties), //Skips 4th descriptor
+	}, null, g_font,'Removes duplicates according to equal ' + enumeratePropertiesValues(newButtonsProperties, prefix, void(0), void(0), 4), prefix, newButtonsProperties, chars.duplicates, _gdiFont('FontAwesome', 10)), //Skips 4th descriptor
 	
 	RemoveDuplicatesV3: new SimpleButton(calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? true : false).x, calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? false : true).y, buttonCoordinates.w, buttonCoordinates.h, 'Filter playlist', function () {
 		let t0 = Date.now();
@@ -62,7 +62,7 @@ var newButtons = {
         do_remove_duplicatesV3(null, null, sortInput, nAllowed);
 		t1 = Date.now();
 		console.log("Call to do_remove_duplicatesV3 took " + (t1 - t0) + " milliseconds.");
-	}, null, g_font,'Filter playlist according to equal ' + enumeratePropertiesValues(newButtonsProperties, prefix, void(0), void(0), 4) + ' and allowing ' + getPropertiesValues(newButtonsProperties,prefix)[3] + ' duplicates', prefix, newButtonsProperties), // Changes a bit the tooltip to show duplicates number separated
+	}, null, g_font,'Filter playlist according to equal ' + enumeratePropertiesValues(newButtonsProperties, prefix, void(0), void(0), 4) + ' and allowing ' + getPropertiesValues(newButtonsProperties,prefix)[3] + ' duplicates', prefix, newButtonsProperties, chars.filter, _gdiFont('FontAwesome', 10)), // Changes a bit the tooltip to show duplicates number separated
 };
 // Check if the button list already has the same button ID
 for (var buttonName in newButtons) {