+++ buttons/buttons_search_same_by.js
@@ -97,7 +97,7 @@ newButtonsProperties['playlistName'].push({func: isString}, newButtonsProperties
 setProperties(newButtonsProperties, prefix); //This sets all the panel properties at once
 
 // we change the default coordinates here to accommodate text
-if (buttonOrientation === 'x') {buttonCoordinates.w += 15;}
+if (buttonOrientation === 'x') {buttonCoordinates.w += 25;}
 
 var newButtons = {
     SearchSameBy: new SimpleButton(calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? true : false).x, calcNextButtonCoordinates(buttonCoordinates, buttonOrientation, buttonOrientation === 'x' ? false : true).y, buttonCoordinates.w, buttonCoordinates.h, 'Search Same By...', function () {
@@ -110,7 +110,7 @@ var newButtons = {
         do_search_same_by(args);
 		t1 = Date.now();
 		console.log("Call to do_search_same_by took " + (t1 - t0) + " milliseconds.");
-	}, null, g_font,'Random playlist matching ' + getPropertiesValues(newButtonsProperties, prefix)[3] +  ' of the current selected track', prefix, newButtonsProperties),
+	}, null, g_font,'Random playlist matching ' + getPropertiesValues(newButtonsProperties, prefix)[3] +  ' of the current selected track', prefix, newButtonsProperties, chars.link, _gdiFont('FontAwesome', 10)),
 };
 // Check if the button list already has the same button ID
 for (var buttonName in newButtons) {